Updated how received requests are fetched and OPCON outsource
Fetches received requests based on if the user is in the request recipients or if the user is in the requests respondents. Additionally, when a Supervisor requests that someone join his team, then both the nearest superior commander of the supervisor, and the nearest superior commander of the new team member must approve.